transaction_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# SQLAlchemy + async driver
SQLAlchemy==2.0.35
asyncpg==0.29.0

# Alembic (opsiyonel, migration için)
alembic==1.13.2

# HTTP client (ileride account service ile iletişim için lazım olacak)
httpx==0.27.0

# Pydantic
pydantic==2.9.2
pydantic-settings==2.4.0

# Python-dotenv (çevre değişkenleri için)
python-dotenv==1.0.1
aio-pika==9.4.1
python-jose[cryptography]
